Use built-in draw() for route partials within a single plugin
