+++ plugins/ShinyCMS/.rubocop.yml
@@ -45,23 +45,17 @@ Bundler/GemVersion:
 # These are mostly expected to contain long blocks due to their DSLs
 Metrics/BlockLength:
   Exclude:
+    # Main app
     - ../../Gemfile
     - ../../config/environments/*
-    - ../Shiny*/config/routes.rb
-    - ../Shiny*/shiny*.gemspec
-    - ../Shiny*/spec/**/*.rb
+    # All plugins
+    - ../*/config/routes.rb
+    - ../*/shiny*.gemspec
+    - ../*/spec/**/*.rb
+    # Core plugin
     - app/models/concerns/shinycms/*.rb
     - app/public/models/concerns/shinycms/*.rb
-    - app/public/lib/shinycms/routes/*.rb
-
-# Likewise not a concern
-Metrics/MethodLength:
-  Exclude:
-    - app/public/lib/shinycms/routes/*.rb
-# Blarg.
-Metrics/AbcSize:
-  Exclude:
-    - app/public/lib/shinycms/routes/*.rb
+    - config/routes/*.rb
 
 # Spec files with I18n.t calls tend to have long lines. I can live with that.
 Layout/LineLength:
